Move headers to top level
