--- include/index.html
@@ -1,86 +1,0 @@
-<!DOCTYPE html>
-
-<meta charset="UTF-8">
-<meta name="robots" content="index,follow" />
-<link rel=icon href="favicon.ico" sizes="any">
-
-<script>
-// Refresh the page on the first second of each minute
-onload = function(){
-  var seconds = 60 - (new Date()).getSeconds();
-  setTimeout(function() {
-    window.location.reload();
-  }, seconds * 1000);
-}
-</script>
-
-<style>
-body {
-  font-family: sans-serif;
-  background-color: #234;
-  color: white;
-  padding: 40px;
-}
-a:link, a:visited { color: lightblue; }
-h1 {
-  font-size: 100%;
-  text-transform: uppercase;
-  color: darkorange;
-}
-p.disclaimer { max-width: 700px; }
-</style>
-
-<title>Have A Nice Day Trader</title>
-<h1>Have A Nice Day Trader <small>DATE</small></h1>
-<p class="disclaimer">
-History is no indicator of future performance. Don't invest what you can't
-afford to lose.
-</p>
-
-<p class="disclaimer">
-24 hours of prices are fetched for SYMBOLS tradable coins from <a
-   href="https://www.cryptocompare.com/api/">CryptoCompare</a> at a rate of
- BATCH per minute, therefore it takes MINUTES minutes to process the full set of
- coins. A library of strategies runs over each batch and a position is created
- if a strategy returns positively. Positions are closed if the return exceeds
-   SELL_THRESHOLD&nbsp;%, falls below CUT&nbsp;% or 24 hours have elapsed since
-     creation. See the code on <a
-   href="http://github.com/deanturpin/handt/">GitHub</a> or <a
-   href="https://github.com/deanturpin/handt/issues/new"" target="___">raise an
-   issue</a>.
-</p>
-
-<p>STATS coins processed since 16&nbsp;April&nbsp;2018.</p>
-
-<p>
-<a href="https://travis-ci.org/deanturpin/handt" target="___">
-  <img src="https://travis-ci.org/deanturpin/handt.svg?branch=master"/></a>
-<tt>GITLOG</tt>
-</p>
-
-<h1>Strategy summary</h1>
-<p>POSITIONS</p>
-<pre>
-ALL COINS               POS     min     mean    max
-ALLCOINS_STRATEGY
-
-COINBASE ONLY           POS     min     mean    max
-COINBASE_STRATEGY
-</pre>
-
-<h1>Coinbase open positions</h1>
-<pre>
-SYMBOL  %       STRATEGY                BUY     SPOT    SELL IN (HOURS)
-COINBASE_OPEN
-</pre>
-
-<h1>Coin performance</h1>
-<pre>
-Symbol  Pos     min     mean    max
-COIN_PERFORMANCE
-</pre>
-
-<h1>Unit test</h1>
-<pre>
-UNIT_TEST
-</pre>
