+++ template.md
@@ -3,6 +3,9 @@
 [![Build
 Status](https://travis-ci.org/deanturpin/handt.svg?branch=master)](https://travis-ci.org/deanturpin/handt)
 
+History is no indicator of future performance. Don't invest what you can't
+afford to lose.
+
 **Have A Nice Day Trader** is an algorithmic trading platform. Prices are
 requested for all currency pairs traded by Binance, a library of strategies is
 then back-tested over the prices and a strategy summary is produced below. See
